requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
accelerate==0.17.1
basicsr==1.4.2
decord==0.6.0
diffusers==0.16.1
einops==0.3.0
ftfy
gradio==3.20.1
imageio==2.9.0
kornia==0.6
langchain==0.0.101
modelscope==1.4.2
moviepy==1.0.3
omegaconf==2.3.0
openai
opencv-python
open_clip_torch
oss2
pandas==2.0.0
Pillow==9.5.0
PyYAML==6.0
ruamel_yaml
timm==0.4.9
tomesd
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.0
transformers==4.28.0
git+https://github.com/suno-ai/bark.git